requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
coverage==7.2.2
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
